Fix no-use-before-define lint error in tutorial-panel.tsx
